feat(core): add ContextService for context generation
