+++ chapter_generation/__init__.py
@@ -1,5 +1,6 @@
 """Utilities for orchestrating chapter generation steps."""
 
+from .context_service import ContextService
 from .drafting_service import DraftingService, DraftResult
 from .evaluation_service import EvaluationCycleResult, EvaluationService
 from .finalization_service import FinalizationService
@@ -13,6 +14,7 @@ __all__ = [
     "DraftResult",
     "EvaluationService",
     "EvaluationCycleResult",
+    "ContextService",
     "RevisionService",
     "RevisionResult",
     "FinalizationService",
