+++ tests/test_orchestrator_private_methods.py
@@ -120,7 +120,8 @@ async def test_prepare_prerequisites_uses_plan(orchestrator, monkeypatch):
         AsyncMock(return_value=([], {})),
     )
     monkeypatch.setattr(
-        "orchestration.nana_orchestrator.generate_hybrid_chapter_context_logic",
+        orchestrator.context_service,
+        "build_hybrid_context",
         AsyncMock(side_effect=fake_context),
     )
